requirements.txt
wxPython>=4.2.1
Pillow>=10.0.0
ImageHash>=4.3.1